requirements.txt
beautifulsoup4
requests
tqdm
pandas
geopy.geocoders
geopandas
time
shapely.geometry
matplotlib.colors
